pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dea-burn-severity"
version = "0.1.0"
description = "CLI for DEA Sentinel-2 burn severity processing"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Digital Earth Australia", email = "dea@ga.gov.au" }]
license = { text = "Apache-2.0" }
dependencies = [
    "click>=8.1,<9",
    "datacube",
    "geopandas",
    "numpy",
    "pandas",
    "xarray",
    "PyYAML",
    "s3fs",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3,<9",
    "pytest-mock>=3.14,<4",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["dea_burn_severity"]
include-package-data = true

[tool.setuptools.package-data]
dea_burn_severity = ["config/*.yaml"]

[project.scripts]
dea-burn-severity = "dea_burn_severity.cli:cli"
